chore: rename com.zextras packages
* use com.zextras.mailbox
