+++ store/conf/web.xml
@@ -26,7 +26,7 @@ SPDX-License-Identifier: AGPL-3.0-only
   </filter-mapping>
 
   <listener>
-    <listener-class>com.zextras.jetty.mailbox.GuiceMailboxServletConfig</listener-class>
+    <listener-class>com.zextras.mailbox.servlet.GuiceMailboxServletConfig</listener-class>
   </listener>
   <filter>
     <filter-name>DoSFilter</filter-name>
